Updated components to work, added unfinished header bar
