+++ client/App.js
@@ -2,10 +2,12 @@ import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, View } from 'react-native';
 import Example from './features/Example';
 import Profile from './features/Profile';
+import Messages from './features/Messages';
 
 export default function App() {
   return (
-    <Profile />
+    <Messages />
+    // <Profile />
     // <View style={styles.container}>
     //   <Example />
     //   <Text>Open up App.js to start working on your app!</Text>
