+++ client/package.json
@@ -16,8 +16,9 @@
     "expo": "~48.0.5",
     "expo-status-bar": "~1.4.4",
     "react": "18.2.0",
-    "react-native": "0.71.3",
+    "react-native": "^0.71.3",
     "react-native-config": "^1.5.0",
+    "react-native-gesture-handler": "^2.9.0",
     "react-native-web": "~0.18.10"
   },
   "devDependencies": {
